requirements.txt
streamlit==1.22.0
plotly==5.18.0
yfinance==0.2.36
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
loguru==0.7.2
pydantic==2.3.0
mpyc==0.10
requests==2.31.0
python-dateutil
python-dotenv
aiohttp
asyncio
pydantic-settings
fastapi
uvicorn
pytest
pytest-asyncio
black
mypy
typing-extensions
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
psycopg2-binary
websockets
alpha_vantage==2.3.1
polygon-api-client==1.12.3
